refactor(NodeService): Convert to abstract class and implement getPrevNodeId in subclasses
